This commit refactors the crate to decouple it from diesel_async
This makes it possible to provide custom implementations of `TaskStore` outside of backie (for
example for sync diesel connections and other backend types)

I've put everything diesel_async specific behind a feature flag that is
enabled by default. The public API with that feature flag enabled stays
mostly the same as demonstrated by the example. Only the `enqueue`
function now needs to know about a specific task store to determine the
backend type.
