+++ Cargo.toml
@@ -21,10 +21,14 @@ thiserror = "1"
 uuid = { version = "1.1", features = ["v4", "serde"] }
 async-trait = "0.1"
 futures = "0.3"
-diesel = { version = "2.1", features = ["postgres", "serde_json", "chrono", "uuid"] }
+diesel = { version = "2.1", features = ["serde_json", "chrono", "uuid"] }
 diesel-derive-newtype = "2.0.0-rc.0"
-diesel-async = { version = "0.3", features = ["postgres", "bb8"] }
-tokio = { version = "1.25", features = ["rt", "time", "macros"] }
+diesel-async = { version = "0.3", features = ["postgres", "bb8"], optional = true }
+tokio = { version = "1.25", features = ["rt", "time", "macros", "sync"] }
 
 [dev-dependencies]
 itertools = "0.10"
+
+[features]
+default = ["async_postgres"]
+async_postgres = ["diesel-async"]
