+++ examples/simple_worker/Cargo.toml
@@ -4,12 +4,12 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-backie = { path = "../../" }
+backie = { path = "../../", features = ["async_postgres"] }
 anyhow = "1"
 env_logger = "0.10"
 log = "0.4.0"
 tokio = { version = "1", features = ["full"] }
-diesel-async = { version = "0.2", features = ["postgres", "bb8"] }
-diesel = { version = "2.0", features = ["postgres"] }
+diesel-async = { version = "0.3", features = ["postgres", "bb8"] }
+diesel = { version = "2.1", features = ["postgres"] }
 async-trait = "0.1"
 serde = { version = "1.0", features = ["derive"] }
